refactor: split analytics-core from analytics.ts
