+++ src/browser.ts
@@ -2,7 +2,8 @@ import { getProcessEnv } from './lib/get-process-env'
 import { getCDN } from './lib/parse-cdn'
 
 import fetch from 'unfetch'
-import { Analytics, AnalyticsSettings, InitOptions } from './analytics'
+import { Analytics } from './analytics'
+import type { AnalyticsSettings, InitOptions } from './analytics-core'
 import { Context } from './core/context'
 import { Plan } from './core/events'
 import { Plugin } from './core/plugin'
