+++ src/plugins/ajs-destination/index.ts
@@ -5,7 +5,7 @@ import {
   SegmentEvent,
 } from '@/core/events'
 import { Alias, Facade, Group, Identify, Page, Track } from '@segment/facade'
-import { Analytics, InitOptions } from '../../analytics'
+import type { AnalyticsCore, InitOptions } from '../../analytics-core'
 import { LegacySettings } from '../../browser'
 import { isOffline, isOnline } from '../../core/connection'
 import { Context, ContextCancelation } from '../../core/context'
@@ -109,7 +109,7 @@ export class LegacyDestination implements Plugin {
     return this.onReady ?? Promise.resolve()
   }
 
-  async load(ctx: Context, analyticsInstance: Analytics): Promise<void> {
+  async load(ctx: Context, analyticsInstance: AnalyticsCore): Promise<void> {
     if (this._ready || this.onReady !== undefined) {
       return
     }
@@ -157,7 +157,7 @@ export class LegacyDestination implements Plugin {
     }
   }
 
-  unload(_ctx: Context, _analyticsInstance: Analytics): Promise<void> {
+  unload(_ctx: Context, _analyticsInstance: AnalyticsCore): Promise<void> {
     return unloadIntegration(this.name, this.version)
   }
